+++ src/api/controladorprecios/app/Http/Controllers/CostoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Contractors\IMapper;
 use App\Contractors\Services\ICostosService;
+use App\Enums\AccionsEnums;
 use App\Mappers\CostoMapper;
 use App\Services\CostosService;
 use App\Services\ServicesContainer;
@@ -22,6 +23,9 @@ class CostoController extends BaseController
 
     public function addCosto(Request $request){
         try {
+
+            if(!$this->HasAccionsPermit([AccionsEnums::ADD_PERMIT,AccionsEnums::ADD_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
+            
             $jsonParsed=$this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             $dto=$this->mapper->reverse($jsonParsed);
@@ -35,6 +39,9 @@ class CostoController extends BaseController
 
     public function addCostos(Request $request){
         try {
+
+            if(!$this->HasAccionsPermit([AccionsEnums::ADD_PERMIT,AccionsEnums::ADD_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
+
             $jsonParsed=$this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             if(!is_array($jsonParsed)) new Response($this->stdResponse(false,true,"invalid content"),400);
@@ -52,6 +59,7 @@ class CostoController extends BaseController
 
     public function getCosto($id){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::READ_PERMIT,AccionsEnums::READ_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             return new Response($this->stdResponse(data:$this->service->getCosto($id)));
         } catch (\Throwable $th) {
             return new Response($this->stdResponse(false,true,$th->getMessage()));
@@ -60,6 +68,7 @@ class CostoController extends BaseController
 
     public function updateCosto(Request $request, $id){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::UPDATE_PERMIT,AccionsEnums::UPDATE_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed=$this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             if(empty($id)) return new Response($this->stdResponse(false,true,"invaid contente"),400);
@@ -78,6 +87,7 @@ class CostoController extends BaseController
 
     public function updateCostos(Request $request){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::UPDATE_PERMIT,AccionsEnums::UPDATE_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed=$this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             if(!is_array($jsonParsed)) new Response($this->stdResponse(false,true,"invalid content"),400);
@@ -95,6 +105,7 @@ class CostoController extends BaseController
 
     public function deleteCosto($id){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::DELETE_PERMIT,AccionsEnums::DELETE_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $this->service->deleteCosto($id);
             return new Response($this->stdResponse());
         } catch (\Throwable $th) {
@@ -104,7 +115,7 @@ class CostoController extends BaseController
 
     public function deleteCostos(Request $request){
         try {
-            
+            if(!$this->HasAccionsPermit([AccionsEnums::DELETE_PERMIT,AccionsEnums::DELETE_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed=$this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             if(!is_array($jsonParsed)) new Response($this->stdResponse(false,true,"invalid content"),400);
@@ -120,6 +131,7 @@ class CostoController extends BaseController
 
     public function getCostosByProveedor($proveedorId){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::READ_PERMIT,AccionsEnums::READ_COSTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $data = $this->service->getCostosByProveedor($proveedorId);
             return new Response($this->stdResponse(data:$data));
         } catch (\Throwable $th) {
