+++ src/api/controladorprecios/app/Http/Controllers/ProductoOrganizacionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Contractors\IMapper;
 use App\Contractors\Services\IProductoOrganizacion;
+use App\Enums\AccionsEnums;
 use App\Mappers\ProductoOrganizacionMapper;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -20,6 +21,7 @@ final class ProductoOrganizacionController extends BaseController
 
     public function addOrganzacion($productoId,Request $request){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::ADD_PERMIT,AccionsEnums::ADD_PRODUCTO_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed= json_decode($request->getContent());
             $dto= $this->mapper->reverse($jsonParsed);
             $publicId=$this->productoOrganizacionService->addOrganizacion($productoId,$dto);
@@ -31,6 +33,7 @@ final class ProductoOrganizacionController extends BaseController
 
     public function deleteOrganizacion($pid){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::DELETE_PERMIT,AccionsEnums::DELETE_PRODUCTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $this->productoOrganizacionService->deleteOrganizacion($pid);
             return new Response($this->stdResponse());
         } catch (\Throwable $th) {
@@ -40,6 +43,7 @@ final class ProductoOrganizacionController extends BaseController
 
     public function getOrganizaciones($productoId){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::READ_PERMIT,AccionsEnums::READ_PRODUCTOS_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $data=$this->productoOrganizacionService->getOrganizaciones($productoId);
             return new Response($this->stdResponse(data:$data));
         } catch (\Throwable $th) {
